+++ src/main/java/com/ericversteeg/InventoryTotalRunData.java
@@ -9,9 +9,9 @@ public class InventoryTotalRunData {
 
     // static item prices so that when ItemManager updates, the Profit / Loss value doesn't all of a sudden change
     // this is cleared and repopulated at the start off each new run (after bank) and whenever new items hit the inventory
-    Map<Integer, Integer> itemPrices = new HashMap<>();
-    Map<Integer, Integer> initialItemQtys = new HashMap<>();
-    Map<Integer, Integer> itemQtys = new HashMap<>();
+    Map<Integer, Long> itemPrices = new HashMap<>();
+    Map<Integer, Long> initialItemQtys = new HashMap<>();
+    Map<Integer, Long> itemQtys = new HashMap<>();
 
     LinkedList<String> ignoredItems = new LinkedList<>();
 }
